Split oplog emitters in files (#14917)
